+++ mars/stn/stn.h
@@ -29,7 +29,6 @@
 #include <vector>
 
 #include "mars/comm/autobuffer.h"
-#include "mars/comm/projdef.h"
 
 namespace mars {
 namespace stn {
@@ -470,7 +469,7 @@ class Callback {
     virtual bool MakesureAuthed(const std::string& _host, const std::string& _user_id) = 0;
 
     // 流量统计
-    virtual void TrafficData(ssize_t _send, ssize_t _recv) = 0;
+    virtual void TrafficData(int64_t _send, int64_t _recv) = 0;
 
     // 底层询问上层该host对应的ip列表
     virtual std::vector<std::string> OnNewDns(const std::string& host, bool _longlink_host) = 0;
